server-lbro/requirements.txt
wheel
Django~=3.0.2
djongo~=1.3.3
pymongo~=3.10.1
djangorestframework~=3.11.0
django-cors-headers~=3.2.1
sqlparse==0.2.4
selenium~=3.141.0
elsapy~=0.5.0
pandas~=0.25.3
tika~=1.23
bibtexparser~=1.1.0
tqdm
